% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.commune.R
\name{summary.commune}
\alias{summary.commune}
\title{Répartition des élus par code professionnel}
\usage{
\method{summary}{commune}(x)
}
\arguments{
\item{df}{Un DataFrame contenant les informations des élus municipaux. Il doit inclure une colonne \code{Code.de.la.catégorie.socio.professionnelle}.}
}
\value{
Un graphique en barres affichant le nombre d'élus par code professionnel.
}
\description{
Cette fonction génère un diagramme en barres représentant la distribution des élus municipaux en fonction de leur code professionnel.
}
